default projects merge and improve
default:
add license header
initialize transceiver thread for all available transceivers
add radio thread when transceiver is available
register radio thread with transceiver for available transceivers
clean up imports
clean up Makefile
set default target to native
include all available modules for native target
add -DSCHEDSTATISTICS option (disabled by default)

default-native:
remove default-native project
